규모별 연도별 지진발생 그룹핑 완료
