+++ src/main/java/com/pcwk/ehr/cmn/EQK_DTO.java
@@ -1,46 +1,46 @@
-package com.pcwk.ehr.cmn;
-
-public class EQK_DTO {
-    private String earthquakeYear;
-    private String magnitudeRange;
-    private int eqkCount;
-
-    public EQK_DTO(String earthquakeYear, String magnitudeRange, int eqkCount) {
-        this.earthquakeYear = earthquakeYear;
-        this.magnitudeRange = magnitudeRange;
-        this.eqkCount = eqkCount;
-    }
-
-    public String getEarthquakeYear() {
-        return earthquakeYear;
-    }
-
-    public void setEarthquakeYear(String earthquakeYear) {
-        this.earthquakeYear = earthquakeYear;
-    }
-
-    public String getMagnitudeRange() {
-        return magnitudeRange;
-    }
-
-    public void setMagnitudeRange(String magnitudeRange) {
-        this.magnitudeRange = magnitudeRange;
-    }
-
-    public int getEqkCount() {
-        return eqkCount;
-    }
-
-    public void setEqkCount(int eqkCount) {
-        this.eqkCount = eqkCount;
-    }
-
-    @Override
-    public String toString() {
-        return "EQK_DTO{" +
-                "earthquakeYear='" + earthquakeYear + '\'' +
-                ", magnitudeRange='" + magnitudeRange + '\'' +
-                ", eqkCount=" + eqkCount +
-                '}';
-    }
-}
+//package com.pcwk.ehr.cmn;
+//
+//public class EQK_DTO {
+//    private String earthquakeYear;
+//    private String magnitudeRange;
+//    private int eqkCount;
+//
+//    public EQK_DTO(String earthquakeYear, String magnitudeRange, int eqkCount) {
+//        this.earthquakeYear = earthquakeYear;
+//        this.magnitudeRange = magnitudeRange;
+//        this.eqkCount = eqkCount;
+//    }
+//
+//    public String getEarthquakeYear() {
+//        return earthquakeYear;
+//    }
+//
+//    public void setEarthquakeYear(String earthquakeYear) {
+//        this.earthquakeYear = earthquakeYear;
+//    }
+//
+//    public String getMagnitudeRange() {
+//        return magnitudeRange;
+//    }
+//
+//    public void setMagnitudeRange(String magnitudeRange) {
+//        this.magnitudeRange = magnitudeRange;
+//    }
+//
+//    public int getEqkCount() {
+//        return eqkCount;
+//    }
+//
+//    public void setEqkCount(int eqkCount) {
+//        this.eqkCount = eqkCount;
+//    }
+//
+//    @Override
+//    public String toString() {
+//        return "EQK_DTO{" +
+//                "earthquakeYear='" + earthquakeYear + '\'' +
+//                ", magnitudeRange='" + magnitudeRange + '\'' +
+//                ", eqkCount=" + eqkCount +
+//                '}';
+//    }
+//}
