+++ src/main/java/com/pcwk/ehr/earthquake/controller/EarthquakeController.java
@@ -1,6 +1,5 @@
 package com.pcwk.ehr.earthquake.controller;
 
-import com.pcwk.ehr.cmn.EQK_DTO;
 import com.pcwk.ehr.cmn.SearchVO;
 import com.pcwk.ehr.earthquake.domain.EarthquakeVO;
 import com.pcwk.ehr.earthquake.service.EarthquakeService;
@@ -31,10 +30,15 @@ public class EarthquakeController {
         search.setSearchWord(searchWord);
         search.setSearchDiv(searchDiv);
 
+        SearchVO search1 = new SearchVO();
+        search1.setSearchWord(searchWord);
+
         try {
             List<EarthquakeVO> list = earthquakeService.getDisMes(search);
-            List<EQK_DTO> eqkByYear = earthquakeService.eqkByYear();
+            List<EarthquakeVO> eqkOccurrence = earthquakeService.eqkOccurrence(search1);
+            List<EarthquakeVO> eqkByYear = earthquakeService.eqkByYear();
             model.addAttribute("list", list);
+            model.addAttribute("eqkOccurrence", eqkOccurrence);
             model.addAttribute("eqkByYear", eqkByYear);
             System.out.println(list);
             System.out.println(eqkByYear);
