+++ src/main/java/com/pcwk/ehr/earthquake/mapper/EarthquakeMapper.java
@@ -1,6 +1,5 @@
 package com.pcwk.ehr.earthquake.mapper;
 
-import com.pcwk.ehr.cmn.EQK_DTO;
 import com.pcwk.ehr.cmn.SearchVO;
 import com.pcwk.ehr.disastermessage.domain.DisasterMessageVO;
 import com.pcwk.ehr.earthquake.domain.EarthquakeVO;
@@ -13,5 +12,6 @@ import java.util.List;
 public interface EarthquakeMapper {
 
     List<EarthquakeVO> getEQK (SearchVO searchVO);
-    List<EQK_DTO> eqkByYear ();
+    List<EarthquakeVO> eqkByYear ();
+    List<EarthquakeVO> eqkOccurrence (SearchVO searchVO);
 }
