+++ src/main/java/com/pcwk/ehr/earthquake/service/EarthquakeService.java
@@ -1,6 +1,5 @@
 package com.pcwk.ehr.earthquake.service;
 
-import com.pcwk.ehr.cmn.EQK_DTO;
 import com.pcwk.ehr.cmn.SearchVO;
 import com.pcwk.ehr.earthquake.domain.EarthquakeVO;
 
@@ -10,5 +9,6 @@ import java.util.List;
 public interface EarthquakeService {
 
     List<EarthquakeVO> getDisMes(SearchVO searchVO) throws Exception;
-    List<EQK_DTO> eqkByYear ();
+    List<EarthquakeVO> eqkByYear ();
+    List<EarthquakeVO> eqkOccurrence (SearchVO searchVO) throws Exception;
 }
